+++ 2019/game/game_test.go
@@ -32,7 +32,7 @@ var _ = Describe("Game", func() {
 	It("can win the game", func() {
 		g.Pay()
 		score := g.Run()
-		Expect(score).To(Equal(int64(19297)))
+		Expect(score).To(Equal(19297))
 	})
 
 })
